Continue refactoring game worker modules
